smyrna: use a 'double' for 'nodeScale'
This is another step towards removing intermediate precision loss in Smyrna.
Squashes a -Wfloat-conversion warning.

+++ cmd/smyrna/smyrnadefs.h
@@ -334,7 +334,7 @@ enum {SCHEMACOUNT = 4}; // number of colors
 	int keyVal; ///< depressed key or 0 if none
 	mouse_action_t* mouse_actions; ///< customizable mouse interaction list
 	size_t mouse_action_count;
-	float nodeScale;
+	double nodeScale;
 	int guiMode;
 	char* optArg;
